refactor: create base class for all dtos with links
