+++ src/CitMovie.Api/GlobalUsings.cs
@@ -2,3 +2,4 @@ global using Microsoft.AspNetCore.Authorization;
 global using Microsoft.AspNetCore.Mvc;
 global using CitMovie.Business;
 global using CitMovie.Models.DataTransferObjects;
+global using CitMovie.Models;
