+++ src/CitMovie.Business/GlobalUsings.cs
@@ -2,4 +2,5 @@ global using Microsoft.Extensions.Options;
 global using CitMovie.Data;
 global using CitMovie.Models.DataTransferObjects;
 global using CitMovie.Models.DomainObjects;
+global using CitMovie.Models;
 global using AutoMapper;
