+++ src/CitMovie.Models/DataTransferObjects/BaseDto.cs
@@ -1,0 +1,4 @@
+namespace CitMovie.Models.DataTransferObjects;
+public class BaseResult {
+    public List<Link> Links { get; set; } = [];
+}
