+++ src/CitMovie.Models/DataTransferObjects/CompletedDto.cs
@@ -1,6 +1,6 @@
 namespace CitMovie.Models.DataTransferObjects;
 
-public class CompletedDto
+public class CompletedDto : BaseResult
 {
     public int CompletedId { get; set; }
     public int UserId { get; set; }
@@ -8,5 +8,13 @@ public class CompletedDto
     public DateTime? CompletedDate { get; set; }
     public int Rewatchability { get; set; }
     public string? Note { get; set; }
-    public List<Link> Links { get; set; } = new List<Link>();
+}
+
+public class CreateCompletedDto
+{
+    public int UserId { get; set; }
+    public int MediaId { get; set; }
+    public DateTime? CompletedDate { get; set; } = DateTime.Now;
+    public int Rewatchability { get; set; }
+    public string? Note { get; set; }
 }
