--- src/CitMovie.Models/DataTransferObjects/CreateBookmarkDto.cs
@@ -1,11 +1,0 @@
-namespace CitMovie.Models.DataTransferObjects
-{
-    public class CreateBookmarkDto
-    {
-        public int UserId { get; set; }
-        public int MediaId { get; set; }
-        public string MediaTitle { get; set; }
-        public string? Note { get; set; }
-    }
-
-}
