+++ src/CitMovie.Models/DataTransferObjects/FollowDto.cs
@@ -1,11 +1,10 @@
 namespace CitMovie.Models.DataTransferObjects;
 
-public class FollowResult
+public class FollowResult : BaseResult
 {
     public int FollowingId { get; set; }
     public int PersonId { get; set; }
     public DateTime FollowedSince { get; set; }
-    public List<Link> Links { get; set; } = new List<Link>();
 };
 
 public class FollowCreateRequest
