+++ src/CitMovie.Models/DataTransferObjects/MediaDto.cs
@@ -2,16 +2,15 @@ using System.Text.Json.Serialization;
 
 namespace CitMovie.Models.DataTransferObjects;
 
-public class MediaBasicResult {
+public class MediaBasicResult : BaseResult {
     public required int Id { get; set; }
     public required string Type { get; set; }
     public required string Title { get; set; }
     public DateTime? ReleaseDate { get; set; }
     public string? PosterUri { get; set; }
-    public List<Link> Links { get; set; } = new List<Link>();
 }
 
-public record MediaResult {
+public class MediaResult : BaseResult {
     public required int Id { get; set; }
     public required string Type { get; set; }
     public required string Title { get; set; }
@@ -27,7 +26,6 @@ public record MediaResult {
     public List<ScoreResult> Scores { get; set; } = [];
     public List<MediaProductionCompanyResult> ProductionCompanies { get; set; } = [];
     public List<CountryResult> ProductionCountries { get; set; } = [];
-    public List<Link> Links { get; set; } = new List<Link>();
 
     public class ScoreResult {
         public required string Source { get; set; }
