+++ src/CitMovie.Models/DataTransferObjects/PersonDto.cs
@@ -1,6 +1,6 @@
 namespace CitMovie.Models.DataTransferObjects;
 
-public record PersonResult
+public class PersonResult : BaseResult
 {
     public int Id { get; set; }
     public string Name { get; set; }
@@ -9,9 +9,8 @@ public record PersonResult
     public decimal? NameRating { get; set; }
     public DateTime? BirthDate { get; set; }
     public DateTime? DeathDate { get; set; }
-    public List<Link> Links { get; set; } = new List<Link>();
 
-    public record MediaResult
+    public class MediaResult : BaseResult
     {
         public int MediaId { get; set; }
         public string Title { get; set; }
@@ -20,16 +19,12 @@ public record PersonResult
         public string? Description { get; set; }
         public string? Awards { get; set; }
         public int BoxOffice { get; set; }
-        public List<Link> Links { get; set; } = new List<Link>();
     }
 }
 
-
-
-public record CoActorResult
+public class CoActorResult : BaseResult
 {
     public string Id { get; set; }
     public string ActorName { get; set; }
     public int Frequency { get; set; }
-    public List<Link> Links { get; set; } = new List<Link>();
 }
