+++ src/CitMovie.Models/DataTransferObjects/SearchHistoryDto.cs
@@ -1,12 +1,10 @@
 namespace CitMovie.Models.DataTransferObjects;
 
 
-public record SearchHistoryResult
+public class SearchHistoryResult : BaseResult
 {
     public int SearchHistoryId { get; set; }
     public int UserId { get; set; }
     public string SearchText { get; set; }
     public string Type { get; set; }
-    public List<Link> Links { get; set; } = new List<Link>();
 }
-
