+++ src/CitMovie.Models/DataTransferObjects/UserScoreDto.cs
@@ -1,18 +1,16 @@
-namespace CitMovie.Models.DataTransferObjects
-{
-    public record UserScoreResult
-    {
-        public int UserId { get; set; }
-        public int MediaId { get; set; }
-        public int Score { get; set; }
-        public string? ReviewText { get; set; }
-        public List<Link> Links { get; set; } = new List<Link>();
-    }
+namespace CitMovie.Models.DataTransferObjects;
 
-    public record UserScoreCreateRequest
-    {
-        public string ImdbId { get; set; }
-        public int Score { get; set; }
-        public string? ReviewText { get; set; }
-    }
+public class UserScoreResult : BaseResult
+{
+    public int UserId { get; set; }
+    public int MediaId { get; set; }
+    public int Score { get; set; }
+    public string? ReviewText { get; set; }
+}
+
+public record UserScoreCreateRequest
+{
+    public string ImdbId { get; set; }
+    public int Score { get; set; }
+    public string? ReviewText { get; set; }
 }
